+++ dialect/mysql/dialect/clauses.go
@@ -28,15 +28,3 @@ type Set struct {
 func (s Set) WriteSQL(ctx context.Context, w io.Writer, d bob.Dialect, start int) ([]any, error) {
 	return bob.Express(ctx, w, d, start, expr.OP("=", expr.Quote(s.Col), s.Val))
 }
-
-type partitions struct {
-	partitions []string
-}
-
-func (h *partitions) AppendPartition(partitions ...string) {
-	h.partitions = append(h.partitions, partitions...)
-}
-
-func (h partitions) WriteSQL(ctx context.Context, w io.Writer, d bob.Dialect, start int) ([]any, error) {
-	return bob.ExpressSlice(ctx, w, d, start, h.partitions, "PARTITION (", ", ", ")")
-}