+++ dialect/psql/dialect/mods.go
@@ -37,19 +37,17 @@ type fromable interface {
 }
 
 func From[Q fromable](table any) FromChain[Q] {
-	return FromChain[Q](func() clause.From {
-		return clause.From{
-			Table: table,
-		}
+	return FromChain[Q](func() clause.TableRef {
+		return clause.TableRef{Expression: table}
 	})
 }
 
-type FromChain[Q fromable] func() clause.From
+type FromChain[Q fromable] func() clause.TableRef
 
 func (f FromChain[Q]) Apply(q Q) {
 	from := f()
 
-	q.SetTable(from.Table)
+	q.SetTable(from.Expression)
 	if from.Alias != "" {
 		q.SetTableAlias(from.Alias, from.Columns...)
 	}
@@ -64,7 +62,7 @@ func (f FromChain[Q]) As(alias string, columns ...string) FromChain[Q] {
 	fr.Alias = alias
 	fr.Columns = columns
 
-	return FromChain[Q](func() clause.From {
+	return FromChain[Q](func() clause.TableRef {
 		return fr
 	})
 }
@@ -73,7 +71,7 @@ func (f FromChain[Q]) Only() FromChain[Q] {
 	fr := f()
 	fr.Only = true
 
-	return FromChain[Q](func() clause.From {
+	return FromChain[Q](func() clause.TableRef {
 		return fr
 	})
 }
@@ -82,7 +80,7 @@ func (f FromChain[Q]) Lateral() FromChain[Q] {
 	fr := f()
 	fr.Lateral = true
 
-	return FromChain[Q](func() clause.From {
+	return FromChain[Q](func() clause.TableRef {
 		return fr
 	})
 }
@@ -91,7 +89,7 @@ func (f FromChain[Q]) WithOrdinality() FromChain[Q] {
 	fr := f()
 	fr.WithOrdinality = true
 
-	return FromChain[Q](func() clause.From {
+	return FromChain[Q](func() clause.TableRef {
 		return fr
 	})
 }
@@ -102,7 +100,7 @@ func Join[Q Joinable](typ string, e any) JoinChain[Q] {
 	return JoinChain[Q](func() clause.Join {
 		return clause.Join{
 			Type: typ,
-			To:   clause.From{Table: e},
+			To:   clause.TableRef{Expression: e},
 		}
 	})
 }
@@ -127,7 +125,7 @@ func CrossJoin[Q Joinable](e any) CrossJoinChain[Q] {
 	return CrossJoinChain[Q](func() clause.Join {
 		return clause.Join{
 			Type: clause.CrossJoin,
-			To:   clause.From{Table: e},
+			To:   clause.TableRef{Expression: e},
 		}
 	})
 }