+++ dialect/psql/dialect/select.go
@@ -14,7 +14,7 @@ type SelectQuery struct {
 	clause.With
 	clause.SelectList
 	Distinct
-	clause.From
+	clause.TableRef
 	clause.Where
 	clause.GroupBy
 	clause.Having
@@ -60,7 +60,7 @@ func (s SelectQuery) WriteSQL(ctx context.Context, w io.Writer, d bob.Dialect, s
 	}
 	args = append(args, selArgs...)
 
-	fromArgs, err := bob.ExpressIf(ctx, w, d, start+len(args), s.From, s.From.Table != nil, "\nFROM ", "")
+	fromArgs, err := bob.ExpressIf(ctx, w, d, start+len(args), s.TableRef, s.TableRef.Expression != nil, "\nFROM ", "")
 	if err != nil {
 		return nil, err
 	}