+++ dialect/psql/dialect/update.go
@@ -12,10 +12,10 @@ import (
 // https://www.postgresql.org/docs/current/sql-update.html
 type UpdateQuery struct {
 	clause.With
-	Only bool
-	clause.Table
+	Only  bool
+	Table clause.TableRef
 	clause.Set
-	clause.From
+	clause.TableRef
 	clause.Where
 	clause.Returning
 	bob.Load
@@ -56,8 +56,8 @@ func (u UpdateQuery) WriteSQL(ctx context.Context, w io.Writer, d bob.Dialect, s
 	}
 	args = append(args, setArgs...)
 
-	fromArgs, err := bob.ExpressIf(ctx, w, d, start+len(args), u.From,
-		u.From.Table != nil, "\nFROM ", "")
+	fromArgs, err := bob.ExpressIf(ctx, w, d, start+len(args), u.TableRef,
+		u.TableRef.Expression != nil, "\nFROM ", "")
 	if err != nil {
 		return nil, err
 	}