+++ dialect/sqlite/im/qm.go
@@ -39,7 +39,7 @@ func OrRollback() bob.Mod[*dialect.InsertQuery] {
 
 func Into(name any, columns ...string) bob.Mod[*dialect.InsertQuery] {
 	return bob.ModFunc[*dialect.InsertQuery](func(i *dialect.InsertQuery) {
-		i.Table = clause.Table{
+		i.TableRef = clause.TableRef{
 			Expression: name,
 			Columns:    columns,
 		}
@@ -48,7 +48,7 @@ func Into(name any, columns ...string) bob.Mod[*dialect.InsertQuery] {
 
 func IntoAs(name any, alias string, columns ...string) bob.Mod[*dialect.InsertQuery] {
 	return bob.ModFunc[*dialect.InsertQuery](func(i *dialect.InsertQuery) {
-		i.Table = clause.Table{
+		i.TableRef = clause.TableRef{
 			Expression: name,
 			Alias:      alias,
 			Columns:    columns,