+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added the `SeparatePackageForTests` attribute to generation outputs to indicate if generated tests should have their own package.
 - Added the `RandomColumnNotNull` mod to factories to generate random values for non-nullable columns without the chance of being null.
 - Added the `WithOneRelations` mod to factories to include every relation for a factory template.
+- Added `clause.TableRef` which merges `clause.Table` and `clause.From` since they had overlapping functionality.
 
 ### Changed
 
@@ -32,6 +33,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   -- After
   SELECT * FROM pilots WHERE jet_id IN (SELECT unnest(CAST($1 AS integer[])); -- Parameters are always 1
   ```
+
+### Removed
+
+- Removed `clause.Table` and `clause.From`, and merge into `clause.TableRef` since they had overlapping functionality.
 
 ### Fixed